Upgrade to Fastify v3
