--- test/types/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "module": "commonjs",
-    "noEmit": true,
-    "strict": true
-  },
-  "files": ["./index.ts"]
-}
